SerialDiscovery: move into the discovery folder
