+++ go/internal/cli/providers/microwendy.go
@@ -55,7 +55,7 @@ func (p *MicroWendyProvider) CheckRequirements(ctx context.Context) error {
 }
 
 func (p *MicroWendyProvider) DiscoverDevices(ctx context.Context) ([]models.ExternalDevice, error) {
-	sd := liteclient.GetSerialDiscovery()
+	sd := discovery.GetSerialDiscovery()
 	sd.StartScan(0)
 
 	services, err := discovery.BrowseMDNSServices(ctx, microWendyServiceType, 3*time.Second)
